+++ src/main/scala/gpdviz/model/model.scala
@@ -19,7 +19,7 @@ case class DataStream(strid:    String,
                       description:  Option[String] = None,
                       mapStyle:     Option[Map[String, JsValue]] = None,
                       zOrder:       Int = 0,
-                      variables:    Option[List[String]] = None,
+                      variables:    Option[JsObject] = None,
                       chartStyle:   Option[JsObject] = None,
                       observations: Option[Map[String, List[ObsData]]] = None,
                       obs:          Option[List[DataObs]] = None